Updated project with folders and script updates
I added some folders to help better separate all of the miscellaneous files that will be floating around in this project. It is all named in such a fashion that my xorg_build_scripts can be added easily to this repo at a future date. I also added a nice .gitignore file to keep things even more tidy.