Added new build script system
While I originally thought that making a single modular script for building tools would be useful, in this case, it was not. Luckily, I came across the project 'LFS---by-bash-scripts' by a GitHub user name 'zajca', who inspired me to build smaller scripts for each package, and then using the source command in a main build script, build them all. On top of this, this does not mean a single modular script couldn't be used either. But these singular scripts look a lot less messy, even if they do include a moderate amount of additional lines. But I think they pay off in the end. So really, this is the way the build scripts are going here on out. And thanks again zacja!
                                ~ Love, Ted M.