Finalized toolchain build scripts
Over the course of the past few days, I have added the files necessary to build the toolchain as specified by the March 6th version of Linux from Scratch Systemd edition. build_toolchain has two main parts: part 1 prepares the directory and sets up a user, and another which is meant to be run by the user created in part 1, which builds the actual toolchain. I might soon add a third part to just mop everything up a little. But anyways, other than this, the other update files were the checksum_update and get_packages scripts, which were just updated to accomidate the new base/toolchain downloads. Of course, many of the sub-scripts were updated, given they are a large part of the debugging process.

Very soon I would like to add the scripts for the base system. These are coming soon. Until next time.
~ Love, Ted M.