Updated scripts to build for RPI3
I recently attempted to build the basic toolchain on my Raspberry Pi 3, and ran into several build errors. While some of these errors allowed me to improve the script, I ran into a major error which I was unsure how to fix. Luckily, my fellow Raspberry Pi user Mogwai (from the PiLFS project) had the fixes for my errors in his own build script. So, I added his small additions to my build scripts for GCC and GLibc. I have yet to test them, but they have been added to this commit nonetheless so that I can pull them onto my Pi.

This also includes the addition of a configuration script for the toolchain build, which can be found at 'toolchain/toolchain.cfg'. This allows for a user to not have to dig around inside of the script and making edits. Instead, they can now just edit this simple file. It is formatted in a "Key: Value" typed fashion, so that it can be read by a simple grep/awk command. Finally, I added a couple Pi specific files to the base.wget file, including a patch, the Pi Bootloader source, and the custom Raspberry Pi variation of the Linux kernel (which has it's own build script as well). So, everything on the Raspberry Pi front is looking quite nice. Just some happy hacking ahead.

~ Love, Ted M.