Updated build script file organization
With the start of the base build scripts, I decided that the folders holding all of the sub-scripts would be more neat if I kept them all in one folder. So I added a build_scripts directory and updated the main build scripts to still function accordingly (which was extremely easy, given that I used variables so regularly for such directories). I also copied over the toolchain files into the base folder, because their build scripts are very similar. So once again, all is going well.
~ Love, Ted M.