Finished testing build scripts
With this late night commit, I have finally finished testing the build scripts, and the should all now succeed! I have only tested them on my PC, so testing on the Raspberry Pi 3 will likely take place at some point tomorrow afternoon. This commit also includes a bunch of smaller updates as well, like the new and improve checksum_update script, which no longer includes my paths, and only the file names. It also has support for keeping the previous sha256 file as well with a '.old' extension. This project is to be completed by this weekend, and I would say that were are still on target. I'm so proud ;'D

~ Love, Ted M.