Updated scripts to work with RPI3
I ran some tests on my Raspberry Pi 3, and as it turns out, the base build wwould fail towards the end of GCC. It may have just been a small error with the toolchain build, but I decided to mop up some of the main files just to ensure it is not a script issue. I will begin testing again in a few minutes. Wish me luck.

~ Love, Ted M.