Fixes added to extra and xorg groups
This is my first commit coming straight from the Raspberry Pi 3. I have successfully debugged the extra group, plus I am 60% through debugging the xorg group. I am having a little bit of an issue with the luit package, which is an X.Org app. BLFS has a patch for it that needs to be added, but unfortunately I can not copy and paste on the Pi right now. So I am commiting this while I have the chance.