Patch xorg group_build and update desktop
In this commit, I added the patch to the groub_build script in the xorg group as I suggested in the previous commit.

I also added hsetroot to the desktop script. Seeing as how imlib2 is added there, why not allow the user to set the background through hsetroot? Exactly. So I am taking it from a GitHub users repo for now. We'll see how well it works out.

~ Love, Ted M.