Tested and debugged gtk and audio groups
Today, I managed to finish up with the building of the X.Org, desktop, gtk, and audio groups! And all on top of the Raspberry Pi as well. I am really quite proud. All of the changes were quite minor besides the fixing of the gtk wget file and the addition of the xinit package for xorg.

~ Love, Ted M.