Added Firefox build group
This morning I managed to piece together the build script for Firefox. This group includes FFmpeg, which may seem like a stretch for the Firefox group. I agree, and it may be moved to a video group or something of that sort in the future. But for now it all works out just fine. This project is very nearly done, just in time to present. Fingers crossed!

~ Love, Ted M.