Test toolchain build on x86_64
The updated toolchain build is now fully functional on x86_64 systems! This is a step in the right direction. Also, some major packages were version bumped, so some patches were removed. Very good news there.

The user can now generate a configuration file rather than edit it manually.It can pick out a few things by itself, but hopefully it can be fully automated in the future.

The next step is to fix up the base build scripts, which may take a considerable amount of time. Also, the installation steps should be updated. All in all, good progress in the right direction!

~ Love, Ted M.